+++ release/nightly_tests/dataset/image_loader_microbenchmark.py
@@ -2,19 +2,29 @@ import ray
 import torch
 import torchvision
 import os
+import time
 import json
 import tensorflow as tf
 
-from benchmark import (
-    crop_and_flip_image_batch,
-    get_transform,
-    iterate,
-)
 
+DEFAULT_IMAGE_SIZE = 224
 
 # tf.data needs to resize all images to the same size when loading.
 # This is the size of dog.jpg in s3://air-cuj-imagenet-1gb.
 FULL_IMAGE_SIZE = (1213, 1546)
+
+
+def iterate(dataset, label, metrics):
+    start = time.time()
+    it = iter(dataset)
+    num_rows = 0
+    for batch in it:
+        num_rows += len(batch)
+    end = time.time()
+    print(label, end - start, "epoch", i)
+
+    tput = num_rows / (end - start)
+    metrics[label] = tput
 
 
 def build_torch_dataset(
@@ -88,6 +98,36 @@ def tf_crop_and_flip(image_buffer, num_channels=3):
     # Flip to add a little more random distortion in.
     image_buffer = tf.image.random_flip_left_right(image_buffer)
     return image_buffer
+
+
+def get_transform(to_torch_tensor):
+    # Note(swang): This is a different order from tf.data.
+    # torch: decode -> randCrop+resize -> randFlip
+    # tf.data: decode -> randCrop -> randFlip -> resize
+    transform = torchvision.transforms.Compose(
+        [
+            torchvision.transforms.RandomResizedCrop(
+                size=DEFAULT_IMAGE_SIZE,
+                scale=(0.05, 1.0),
+                ratio=(0.75, 1.33),
+            ),
+            torchvision.transforms.RandomHorizontalFlip(),
+        ]
+        + [torchvision.transforms.ToTensor()]
+        if to_torch_tensor
+        else []
+    )
+    return transform
+
+
+def crop_and_flip_image_batch(image_batch):
+    transform = get_transform(False)
+    batch_size, height, width, channels = image_batch["image"].shape
+    tensor_shape = (batch_size, channels, height, width)
+    image_batch["image"] = transform(
+        torch.Tensor(image_batch["image"].reshape(tensor_shape))
+    )
+    return image_batch
 
 
 if __name__ == "__main__":
