+++ release/release_tests.yaml
@@ -6180,11 +6180,11 @@
     byod:
       type: gpu
     cluster_env: app_config.yaml
-    cluster_compute: ../../air_tests/air_benchmarks/compute_cpu_1_aws.yaml
+    cluster_compute: ../../air_tests/air_benchmarks/compute_cpu_4_aws.yaml
 
   run:
     timeout: 600
-    script: python train_benchmark.py --num-workers 1
+    script: python image_train_benchmark.py --num-workers 4
 
   variations:
     - __suffix__: aws
@@ -6193,7 +6193,7 @@
       frequency: manual
       cluster:
         cluster_env: app_config.yaml
-        cluster_compute: ../../air_tests/air_benchmarks/compute_cpu_1_gce.yaml
+        cluster_compute: ../../air_tests/air_benchmarks/compute_cpu_4_gce.yaml
 
 - name: read_images_train_16_cpu
   group: data-tests
@@ -6210,7 +6210,7 @@
 
   run:
     timeout: 600
-    script: python train_benchmark.py --num-workers 16
+    script: python image_train_benchmark.py --num-workers 16
 
   variations:
     - __suffix__: aws
